Allow using XDG_CACHE_HOME on non-linux platforms.
I'd like to be able to use XDG_CACHE_HOME on non-linux platforms (more specifically, macOS).

The XDG directory specification is primarily used on linux, but many tools don't actually care about the platform they're running on as to whether they accept that particular environment variable, so I'd like to propose a small change to `config-path.js` to have an early return of `linux()`, before the platform is taken into account.

This will allow me, (and hopefully others with the same environment variable), to keep their home directories clean of cached files, and localise them all to one place.